+++ web-admin/src/main/java/com/isystk/sample/web/admin/controller/html/post/edit/PostEditForm.java
@@ -1,5 +1,6 @@
 package com.isystk.sample.web.admin.controller.html.post.edit;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import javax.validation.constraints.Digits;
@@ -23,7 +24,7 @@ public class PostEditForm extends BaseForm {
   private static final long serialVersionUID = 7593564324192730932L;
 
   @NotNull
-  Integer userId;
+  BigInteger userId;
 
   @NotNull
   Integer postId;
